Make grub timeout configurable and set to 0
There are situations where it is possible for grub to get stuck waiting
for user input, e.g. if no keyboard is connected or if there is any
other keypress during the grub menu countdown. Changing the timeout to 0
prevents this.

This is also made into a configurable parameter in case someone needs to
use a non-zero value for troubleshooting.

This only applies to the ephemeral ISO.

Change-Id: I34848b7fc56c5698fd3c408bf82ace53e5366d13